+++ lib/movies/pages/movies_page/movies_list.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class MoviesList extends StatelessWidget {
+  const MoviesList({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
