+++ lib/movies/pages/movies_page/settings_button.dart
@@ -1,0 +1,80 @@
+import 'package:flutter/material.dart';
+
+class SettingsButton extends StatelessWidget {
+  const SettingsButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      icon: const Icon(Icons.auto_awesome_outlined),
+      iconSize: 42,
+      onPressed: () async {
+        await showModalBottomSheet<bool>(
+          context: context,
+          shape: const RoundedRectangleBorder(
+            borderRadius: BorderRadius.only(
+              topLeft: Radius.circular(40),
+              topRight: Radius.circular(40),
+            ),
+          ),
+          builder: (BuildContext context) {
+            return const _ViewBottomSheet();
+          },
+        );
+      },
+    );
+  }
+}
+
+class _ViewBottomSheet extends StatelessWidget {
+  const _ViewBottomSheet();
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 200,
+      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+      child: Center(
+        child: Column(
+          children: [
+            Container(
+              width: 30,
+              height: 5,
+              decoration: BoxDecoration(
+                borderRadius: BorderRadius.circular(20),
+                color: Colors.black,
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.symmetric(
+                horizontal: 16,
+                vertical: 10,
+              ),
+            ),
+            ListTile(
+              leading: const Icon(Icons.grid_3x3),
+              title: const Text('グリッド表示に切り替え'),
+              subtitle: const Text(
+                '映画をグリッド表示します',
+              ),
+              onTap: () {
+                Navigator.pop(context);
+              },
+            ),
+            OutlinedButton(
+              style: OutlinedButton.styleFrom(
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(40),
+                ),
+              ),
+              child: const Text('キャンセル'),
+              onPressed: () {
+                Navigator.pop(context, false);
+              },
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
